[#162727556]: Add article validation
