+++ middlewares/articleValidation.js
@@ -1,0 +1,41 @@
+const requiredParams = ['title', 'body', 'description', 'category'];
+
+export default {
+  expectedParamsValidator(req, res, next) {
+    const errorArray = [];
+
+    requiredParams.forEach((param) => {
+      if (!Object.keys(req.body).includes(param)) {
+        errorArray.push(`${param} is required`);
+      }
+    });
+
+    if (!errorArray.length) {
+      return next();
+    }
+
+    const errorMessage = JSON.stringify(errorArray);
+    const error = new Error(errorMessage);
+    error.status = 400;
+    return next(error);
+  },
+
+  nonEmptyParamsValidator(req, res, next) {
+    const errorArray = [];
+
+    requiredParams.forEach((param) => {
+      if (!req.body[param].trim().length) {
+        errorArray.push(`${param} must not be empty`);
+      }
+    });
+
+    if (!errorArray.length) {
+      return next();
+    }
+
+    const errorMessage = JSON.stringify(errorArray);
+    const error = new Error(errorMessage);
+    error.status = 400;
+    return next(error);
+  }
+};
