+++ middlewares/index.js
@@ -2,3 +2,4 @@ export { default as errorHandler } from './errorHandler';
 export { default as userValidations } from './userValidations';
 export { default as notFoundRoute } from './notFoundRoute';
 export { default as verifyToken } from './authentication';
+export { default as articleValidation } from './articleValidation';
