+++ routes/articleRoutes.js
@@ -1,12 +1,19 @@
 import express from 'express';
 import Articles from '../controllers/article';
-import { verifyToken } from '../middlewares';
+import { verifyToken, articleValidation } from '../middlewares';
 
 const router = express.Router();
+
 const { createArticle } = Articles;
+const {
+  expectedParamsValidator,
+  nonEmptyParamsValidator
+} = articleValidation;
 
 router.route('/articles')
   .all(verifyToken)
-  .post(createArticle);
+  .post(expectedParamsValidator,
+    nonEmptyParamsValidator,
+    createArticle);
 
 export default router;
